+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.0] - 2024-09-24
+## [2.0.1] - 2023-09-26
+
+### Fixed
+- Corrected handling of array schemas in the factory function
+- Added warning when `isSample` option is incorrectly used with array schemas
+
+## [2.0.0] - 2024-09-26
 
 ### Changed
 - Refactored `factory` function to use an options object for additional parameters.